cleaning up idl documentation and Miro namspace in idl files moved the constants HORIZON_RANGE, GROUPWISE etc. into the Miro::RangeSensor interface. cleaning up the buttons interface propagating the changes through the code

+++ trunk/src/sparrow/SparrowDevice.h
@@ -30,7 +30,7 @@ namespace Sparrow
   // Hardware specifica
   //--------------------------------------------------------------------------
 
-  static const int NUMBER_OF_BUTTONS = 2;
+  static const unsigned long NUMBER_OF_BUTTONS = 2;
 
   static const int SERVO_FACTOR = 27;
